major refactoring. moved all components into app/components. all imports should be fixed
