+++ src/app/(home)/layout.js
@@ -1,18 +1,18 @@
 "use client";
 
 // React
-import React, {useState} from "react";
+import React, { useState } from "react";
 
 // Material UI
-import {useTheme} from "@mui/material/styles";
+import { useTheme } from "@mui/material/styles";
 import Box from "@mui/material/Box";
 import CssBaseline from "@mui/material/CssBaseline";
 import useMediaQuery from "@mui/material/useMediaQuery";
 
 // Local
 import AppBar from "@/components/AppBar/AppBar"; // Import AppBar component
-import Drawer from "./Drawer"; // Import Drawer component
-import BottomNavigation from "./BottomNavigation";
+import Drawer from "../../components/Drawers/Drawer"; // Import Drawer component
+import BottomNavigation from "../../components/Layouts/BottomNavigation";
 
 // Global Variables
 const drawerWidth = 240;
@@ -27,7 +27,7 @@ const drawerWidthClosed = 100;
  *
  * @returns {ReactElement} The React Element created by this function.
  */
-export default function Layout({children}) {
+export default function Layout({ children }) {
     const theme = useTheme();
     const isMobile = useMediaQuery(theme.breakpoints.down("md")); // Using a media query to determine screen size
     const [open, setOpen] = useState(!isMobile);
@@ -66,9 +66,8 @@ export default function Layout({children}) {
                     p: 3,
                     width: {
                         xs: "100%",
-                        md: `calc(100% - ${
-                            open ? drawerWidth : drawerWidthClosed
-                        }px)`,
+                        md: `calc(100% - ${open ? drawerWidth : drawerWidthClosed
+                            }px)`,
                     },
                     transition: (theme) =>
                         theme.transitions.create("margin", {
@@ -84,7 +83,7 @@ export default function Layout({children}) {
                             theme.transitions.create("margin", {
                                 easing: theme.transitions.easing.easeOut,
                                 duration:
-                                theme.transitions.duration.enteringScreen,
+                                    theme.transitions.duration.enteringScreen,
                             }),
                     }),
                 }}
@@ -98,10 +97,10 @@ export default function Layout({children}) {
                 />
 
                 {/* Children */}
-                <Box sx={{mt: 7}}>{children}</Box>
+                <Box sx={{ mt: 7 }}>{children}</Box>
 
                 {/* Bottom Navigation */}
-                <BottomNavigation/>
+                <BottomNavigation />
             </Box>
         </Box>
     );
