--- src/app/(home)/pathways/components/EditPathwayDialog.js
@@ -1,126 +1,0 @@
-import React from "react";
-import Dialog from "@mui/material/Dialog";
-import {v4 as uuid} from "uuid";
-import DialogTitle from "@mui/material/DialogTitle";
-import DialogContent from "@mui/material/DialogContent";
-import TextField from "@mui/material/TextField";
-import Typography from "@mui/material/Typography";
-import Box from "@mui/material/Box";
-import Accordion from "@mui/material/Accordion";
-import AccordionSummary from "@mui/material/AccordionSummary";
-import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-import IconButton from "@mui/material/IconButton";
-import AddCircleOutlineIcon from "@mui/icons-material/AddCircleOutline";
-import Chip from "@mui/material/Chip";
-import DialogActions from "@mui/material/DialogActions";
-import Button from "@mui/material/Button";
-
-export default function EditPathwayDialog({open, handleClose, handleEditPathway, pathwayFields}) {
-
-    const fields = {
-        id: pathwayFields.id,
-        pathwayTitle: pathwayFields.pathwayTitle,
-        degreeMajor: pathwayFields.degreeMajor,
-        school: pathwayFields.school,
-        yearOfGrad: pathwayFields.yearOfGrad,
-        degreeType: pathwayFields.degreeType
-    }
-
-    return (
-        <>
-        <Dialog
-            open={open}
-            onClose={handleClose}
-            PaperProps={{
-                component: 'form',
-                onSubmit: (e) => {
-                    e.preventDefault();
-                    const formData = new FormData(e.currentTarget);
-                    const formJson = Object.fromEntries(formData.entries());
-                    const newPathwayTitle = formJson.pathwayTitle;
-                    const newDegreeMajor = formJson.degreeMajor;
-                    const newSchool = formJson.school;
-                    const newYearOfGrad = formJson.yearOfGrad;
-                    const newDegreeType = formJson.degreeType;
-
-                    handleEditPathway(fields.id, {newPathwayTitle, newDegreeMajor, newSchool, newYearOfGrad, newDegreeType});
-                    handleClose();
-                }
-
-            }}
-        >
-            <DialogTitle>Edit Pathway: {fields.id}</DialogTitle>
-            <DialogContent>
-                <TextField
-                    autoFocus
-                    required
-                    margin="dense"
-                    id="pathwayTitle"
-                    name="pathwayTitle"
-                    label="Pathway Title"
-                    type="text"
-                    fullWidth
-                    variant="standard"
-                    defaultValue={fields.pathwayTitle}
-                />
-                <TextField
-                    autoFocus
-                    required
-                    margin="dense"
-                    id="degreeMajor"
-                    name="degreeMajor"
-                    label="Degree Major"
-                    type="text"
-                    fullWidth
-                    variant="standard"
-                    defaultValue={fields.degreeMajor}
-                />
-                <TextField
-                    autoFocus
-                    required
-                    margin="dense"
-                    id="school"
-                    name="school"
-                    label="School"
-                    type="text"
-                    fullWidth
-                    variant="standard"
-                    defaultValue={fields.school}
-                />
-                <TextField
-                    autoFocus
-                    required
-                    margin="dense"
-                    id="yearOfGrad"
-                    name="yearOfGrad"
-                    label="Year of Grad"
-                    type="text"
-                    fullWidth
-                    variant="standard"
-                    defaultValue={fields.yearOfGrad}
-                />
-                <TextField
-                    autoFocus
-                    required
-                    margin="dense"
-                    id="degreeType"
-                    name="degreeType"
-                    label="Degree Type"
-                    type="text"
-                    fullWidth
-                    variant="standard"
-                    defaultValue={fields.degreeType}
-                />
-
-
-
-
-            </DialogContent>
-            <DialogActions>
-                <Button onClick={handleClose}>Cancel</Button>
-                <Button type="submit">Edit</Button>
-            </DialogActions>
-        </Dialog>
-</>
-    )
-}
